+++ swift/Sources/LibSignalClient/Net.swift
@@ -27,6 +27,7 @@ public class Net {
 
     /// Creates a new `Net` instance that enables interacting with services in the given Signal environment.
     public init(env: Environment, userAgent: String, remoteConfig: [String: String] = [:]) {
+        self.environment = env
         self.asyncContext = TokioAsyncContext()
         self.connectionManager = ConnectionManager(env: env, userAgent: userAgent, remoteConfig: remoteConfig)
     }
@@ -261,11 +262,17 @@ public class Net {
     /// - Returns:
     ///   An object representing the established, but not active, connection.
     public func connectUnauthenticatedChat() async throws -> UnauthenticatedChatConnection {
-        return try await UnauthenticatedChatConnection(tokioAsyncContext: self.asyncContext, connectionManager: self.connectionManager)
+        return try await UnauthenticatedChatConnection(
+            tokioAsyncContext: self.asyncContext,
+            connectionManager:
+            self.connectionManager,
+            environment: self.environment
+        )
     }
 
     internal var asyncContext: TokioAsyncContext
     internal var connectionManager: ConnectionManager
+    internal let environment: Environment
 }
 
 /// Authentication information used for connecting to CDS and SVR servers.