+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.73.2
 
 - Android: Fix bug which can cause CompletableFuture to be stripped by ProGuard/R8.
+
+- keytrans: Bridge to ffi